+++ src/Utility/Collection.php
@@ -13,13 +13,18 @@
  */
 namespace Origin\Utility;
 
-use ArrayAccess;
 use Iterator;
 use Countable;
 
-class Collection implements ArrayAccess, Iterator, Countable
+class Collection implements Iterator, Countable
 {
+    /**
+     * Holds the Items
+     *
+     * @var array|Origin\Model\Collection
+     */
     protected $items = null;
+
     protected $position = 0;
 
     public function __construct($items)
@@ -615,30 +620,7 @@ class Collection implements ArrayAccess, Iterator, Countable
         }
         return $value;
     }
-    // ArrayAccess
-    public function offsetExists($key)
-    {
-        return array_key_exists($key, $this->items);
-    }
  
-    public function offsetSet($key, $value)
-    {
-        if (is_null($key)) {
-            $this->items[] = $value;
-        } else {
-            $this->items[$key] = $value;
-        }
-    }
- 
-    public function offsetGet($key)
-    {
-        return $this->items[$key];
-    }
-
-    public function offsetUnset($key)
-    {
-        unset($this->items[$key]);
-    }
     // Interable
     public function rewind()
     {